Update Added Deep.BI
Hello, thank you for your consideration! The phrasing I have put in there might have been better in terms of making it clear but the main point is that our platform is built on Druid (on which our UI platform is built as well) as the core with others such as Flink and Cassandra that allow us to augment the way we use Druid.

Thanks in advance!